dev6a/startup: on startup hold SD0 pins via pull ups/downs
on reset all pins are switched to inputs.  While we are starting up,
prior to initilizing the SD0 pins to their proper states, we want
to hold these pins in a reasonable state via pull ups and pull downs.

This only happens for a very short period of time until the pins are
put into their proper state via pins_init.