gps_v4: handle interrupts for SWVER
o turn rx_interrupts off when starting SW_VER request
o turn rx_interrupts on when entering ON from SW_VER

With rx_interrupts on when doing the SW_VER request will fail
because the probe will first get an ACK and then the probe
response.  The SW_VER will get triggered by the ACK and when
the probe comes in the SW_VER will blow up because not
expected.

Don't let rx_interrupts in during a SW_VER request.  This
avoids the problem.