gps/msgbuf: msg_next needs to check for MSG_FILLING too.
fixes gps_panic throw.  If the gps sends enough messages we could
processing messages and then eventually pick up a message that
has been started but not completed.

We need to check not only for an empty queue, but that the 1st
element of the queue has been completed as well.  If not return
NULL so it stops the queue processing.