tagfmtsd: zero the ImageManager directory
when formating a new SD we want to zero the ImageManger
directory.  When ImageManager boots and finds the 1st
sector (ie. the directory) as all zeros, it will to the
initial directory write with an empty directory.

This avoid the tagfmtsd program having to know about
what a directory looks like.