Tagnet: corrected error in handling name.next_element()
The name.next_element() function changes state and it was
being called in every node of the name tree, regardless of
match or not. So a failing node would still call it and
advance the name element to the next, and probably non-existent,
element.

Added the function name.is_last_element(msg) to test if the
current (this) element is the last in the name. It's now used
if a node needs to test if there is more to the name.