collect/sync: fix laying down last SYNC on restart
During a system reboot, a SysReboot.shutdown_flush is signalled.  This
tells various parts of the system to flush any data they may have.

If Collect is working on a buffer (ie. a buffer is ALLOC'd and being filled
in by collect) and there is room, we lay down a SYNC record before telling
the StreamWriter (SSW) to actually to the flush.

However, we build the sync by hand and didn't get it right.  We need to
fill in the current recnum and compute a valid checksum over the whole
record.

o refactor code that fills in recnum.  create finish_record that messes
  with the record number and computes the checksum.

o rework SysReboot.shutdown_flush to use finish_record to fix the bad
  SYNC we were laying down.