tagdump: update populators to include object name
when looking at populator tuples objects display with whatever their
value has been set to, ie. -1 if uninitialized.  It is helpful to see
what their actual object names are.

We add an extra element at the end (doesn't effect the accessor identifier)
that takes care of this when displaying tuples.