collect/sync: add DT_SYNC_FLUSH, update tagdump and mm
when we flush buffers (system_flush), we lay down a last sync
record if possible.  We relabel this SYNC_FLUSH and this tells
tagdump and others that this is the last record in the dblk
stream before a reboot.

The next record is guaranteed to start on the next sector boundary
and will be a reboot record.