gps/mon: make gps restart properly when in MPM (panic/soft_reset)
if a PANIC occurs when the gps is in MPM mode a soft_reset keeps the gps
powered but when we reboot the gps wouldn't respond to SWVER which is part
of the startup state machine.  The gps is in MPM and doesn't respond to
incoming packets until it is woken up via a pulse.

Make the CONFIG (SWVER) state machine compensate by pulsing on even
timeouts.  This is a heuristic that should handle all cases we currently
know about.