panic: add panic_info into panic_info_t structure
PIX needs to be able to grab the panic info easily for display.
keep a copy in panic_hdr_0 (in the panic_info_t structure).

Note: there was a bug that made panic_hdr0_t too big by 4 bytes.
This lost the ph0_checksum.  But it wasn't being checked so fell
through the cracks.  Shrink image_info by 4 additional bytes to
compensate.