tagnet/monitor: forme_task, if not in RECV, panic.
forme_task handles incoming packet that are definitely for us.  There is
existing code that checks for !RECV and processes the packet.

But how can it ever get there?

At some point the Monitor has to be in the RECV state (and the Radio is
RX_ON or beyond, its receiving).

We have a packet (that's why we are in forme_task).  Therefore we have
been in receive state (see above).  The only way out of receive is the
smTimer went off.

If the smTimer went off prior to RadioReceive.receive it will try to put
the Radio into Standby but will get a EBUSY back (busy receiving a packet).
If the smTimer goes off after RadioReceive.receive then it will see that
tagMsgBusy is set and stay in RECV.

So there shouldn't be anyway to get into forme_task with a state other than
RECV.  So for now let it hit the Panic to make sure.