tagnet/mon: patch race condition between radio/monitor
Panic in tagmon_forme_task checks for being not being in recv and would
roll over.  Ie. we got a packet but by the time we started processing our
state had changed to non-recv (ie. the recv window had closed).

This isn't supposed to happen.  The root cause is a race condition between
the interrupt driver and the synchronous monitor.  See code for details.
There are monuments.