TagCore: add Emitter for Influx Database HTTP line protocol
- define new emitter and add to all of the dt records
- distinquish hex values from char strings

This is to allow the Influx emitter to work properly, since the
Influx Database auto-generates the schema. If the first time
a field is encountered it looks like a number, then the schema
is created with expectation of this field being integer. If
a following use of this field has a string, like 'd1' vs '1d',
then there is a conflict in data type.