coretime: add logging points to boot sequence.
add a couple of logging taps to the CoreTime boot sequence.
also remove the CoreTime.verify() (checks R vs. PS) at the end
of dco_sync_task.  This would cause a panic because R and PS
diverge.  For now we let it.  Not clear if R and PS can be kept
sync'd to each other.