tagcore: make resync handle end of data file properly.
when we run off the end of a data file, resync tried to read one
more resync record and the record.set() would fail.

Check for a sync read that is too short.  If so, assume that
we are past the EOF.  --tail will never return a record too short.