gps: handle gsd4e h/w config for 115200.
Tib5 is strapped to come up as 115200/binary.  While switching
over to 9600/binary Tib5 wouldn't configure properly.  Correct
that oversight.

We add two more entries to the probe table for the 9600 TARGET.
115200/OSP goes to 9600/OSP and 115200/NEMA goes to 9600/OSP.
Tib5 is strapped to come up as 115200/OSP.

In addition we change the sirf_factory_clear message to set the port
to default instead of 115200/OSP.