gps: switch to 9600 baud from 115200.
when running at 115200, the interarrival time for each byte is
approx. 86uS.  When running with a 16MHz cpu clock this yields about
250 instructions for handling each byte.  This gets even worse if
DMA is running because of cycle stealing.

We were seeing overruns on the gps port running at 115200.  Changing to
9600 baud fixes this problem.  The downside is slower communcations
by a factor of 10.  But this doesn't look like it hurts anything.