dblk: protect dblk time sync, only set if timesrc < GPS0
If gps time has gone backwards wrt current rtc time (rtc is fast), we would
jam gps time into rtc and reboot.

On the reboot, dblk is read and the last record's time is extracted.  It is
possible for this time to be later than current time (gps time was older
than rtc), and we would use the dblk time for rtc time.

But this is wrong.  The new gps time is the one we want to use.  But the
dblk time gets used because it looks better than the current time.
(Younger).

Make GPS0/GPS times have the highest priority.