sensor/mm6: rework sensor ids to make room for various ACCELn ids.
We use the sensor_id (sns_id) to indicate the data format being
used by the ACCELn sensor.  The LisXdh family can use 8, 10, and
12 bit data depending on the configuration.  Our driver supports
all 3 formats, not at the same time.

We could have used another data cell to indicate the data format.
The down side of that approach is a) the extra data cell and b)
an extra data cell read when processing the data.

The sns_id space is 16 bits so there are plenty of space.  This
keeps the decode very simple.