core/mr: emit_sensor_data_mr should use xdict/mr_emitter scheme.
If a sensor has a simple emission scheme it can provide a dict that
defines what fields are displayed.  The dict keys are the name of the
field and are used to label (if displayed) the column/field.  The
values of the sensor are stored on the key.

Alternatively, if a sensor needs a programatic scheme, the dict function
is left as None and a function is assigned in the SNS_MR_EMITTER field
in the corresponding entry of the sensor dispatch table.