dev7/gps: txrdy_int_enable needs to not lose any interrupts.
for historical reasons, the enable used to call edgeRising which
could generate a spurious interrupt so it had to be cleared.  This is
from when the enable typically would only be called once on boot.

We use enable/disable in multiple places and really just want it
to enable the interrupt.