Collect/SSW: provide mechanism for standalone flush via Collect.
Sometimes when starting up it is handy to use Collect to put entries
in the data store.  But if these are standalone it is possible to
overflow the SSW/SD buffers because the SD writer isn't running yet.

Provide a mechanism to allow run to completion segments to force SSW/SD
buffers to get pushed to the SD.

call Collect.sa_flush() to flush any full buffers out to the SD.

    Collect.sa_flush()
        SSW.start_sa_flush    to set state and prepare for SA write.
                              sets SSW state to FLUSH.

        SSW.sa_flush(flush_all, clear) call to perform actual push.  flush_all
                              when TRUE says also push last ALLOC buffer.
                              clear when TRUE says mark buffer empty after
                              flushing.

    SSW.flush_all -> SSW.flush_sa(flush_all, clear).