ubx: capture upd_sos_clear.
will be removed in next commit.

according to the R24 documentation, on boot after powering back up folling
an UPD-SOS-SAVE, the database will be restore and a UPD-SOS-RESTORED
message is sent.  One should then send UPD-SOS-CLEAR to clear the flash.

We however don't keep the gps powered up.  And having any data
base (almanac) is better than none.  We just keep it.  Prior to powering
down again we will do another UPD-SOS-SAVE which will refresh what is in
the flash.